Progress on Birthdays proj

+++ week-9/rvw-1/birthdays/app.py
@@ -1,4 +1,5 @@
 import os
+import sys
 
 from cs50 import SQL
 from flask import Flask, flash, jsonify, redirect, render_template, request, session
@@ -13,21 +14,21 @@ app.config["TEMPLATES_AUTO_RELOAD"] = True
 db = SQL("sqlite:///birthdays.db")
 
 
-@app.after_request
-def after_request(response):
-    """Ensure responses aren't cached"""
-    response.headers["Cache-Control"] = "no-cache, no-store, must-revalidate"
-    response.headers["Expires"] = 0
-    response.headers["Pragma"] = "no-cache"
-    return response
+# @app.after_request
+# def after_request(response):
+#     """Ensure responses aren't cached"""
+#     response.headers["Cache-Control"] = "no-cache, no-store, must-revalidate"
+#     response.headers["Expires"] = 0
+#     response.headers["Pragma"] = "no-cache"
+#     return response
 
 
-@app.route("/", methods=["GET", "POST"])
+@app.route("/", methods=["GET", "DELETE", "POST", "PUT"])
 def index():
     if request.method == "POST":
 
         # TODO: Add the user's entry into the database
-        name = request.form.get("name")
+        name = request.form.get("birth_name")
         month = request.form.get("birth_month")
         day = request.form.get("birth_day")
 
@@ -39,18 +40,32 @@ def index():
 
         return redirect("/")
 
-    else:
+    elif request.method == "PUT":
+        
+        birthday_id = request.form.get("id")
+        name = request.form.get("birth_name")
+        month = request.form.get("birth_month")
+        day = request.form.get("birth_day")
+        
+        if birthday_id and name and month and day:
+            db.execute(
+                "UPDATE birthdays SET name = ?, month = ?, day = ? WHERE id = ?",
+                name, month, day, birthday_id
+            )
+
+        return redirect("/")
+
+    elif request.method == "DELETE":
+        birthday_id = request.args.get()
+        if birthday_id:
+            db.execute("DELETE FROM birthdays WHERE id = ?", birthday_id)
+
+        return redirect("/")
+
+    elif request.method == "GET":
 
         # Display the entries in the database on index.html
         birthdays = db.execute("SELECT * FROM birthdays")
 
-        return render_template("index.html", birthdays=birthdays)
-
-
-@app.route("/delete", methods=["POST"])
-def delete():
-    birthday_id = request.form.get("birthday_id")
-    if birthday_id:
-        db.execute("DELETE FROM birthdays WHERE id = ?", birthday_id)
-
-    return redirect("/")
+        # return render_template("index.html", birthdays=birthdays)
+        return jsonify(birthdays)
